requirements.txt
slack-bolt==1.18.1
feedparser==6.0.11
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dateutil==2.8.2
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
pyyaml==6.0.1
aiohttp==3.9.1
aiodns==3.1.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0